requirements.txt
requests==2.32.5
beautifulsoup4==4.14.2
PyGithub==2.5.0
PyMuPDF==1.24.14
Pillow==11.0.0
pytesseract==0.3.13
python-dotenv==1.1.1
click==8.3.0
pytest==8.4.2
pytest-cov==7.0.0
